+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Log file moved to platform-specific directory [#140](https://github.com/anlar/tewi/issues/140)
 - Optimize Jackett search (use 'all' endpoint or run indexers in parallel)
 - Use ruff as linter (reformatted whole code base) [#135](https://github.com/anlar/tewi/issues/135)
+- GeoIP module gracefully handles missing geoip2fast dependency
 
 ### Removed